+++ Cesium3DTilesSelection/src/Tileset.cpp
@@ -7,6 +7,7 @@
 #include "Cesium3DTilesSelection/TileID.h"
 #include "Cesium3DTilesSelection/TilesetLoadFailureDetails.h"
 #include "Cesium3DTilesSelection/spdlog-cesium.h"
+#include "QuantizedMeshContent.h"
 #include "TileUtilities.h"
 #include "TilesetLoadIonAssetEndpoint.h"
 #include "TilesetLoadSubtree.h"
@@ -14,6 +15,7 @@
 #include "TilesetLoadTilesetDotJson.h"
 
 #include <CesiumAsync/AsyncSystem.h>
+#include <CesiumAsync/IAssetResponse.h>
 #include <CesiumGeometry/Axis.h>
 #include <CesiumGeometry/TileAvailabilityFlags.h>
 #include <CesiumGeospatial/Cartographic.h>
@@ -334,6 +336,11 @@ Tileset::updateView(const std::vector<ViewState>& frustums) {
           pCreditSystem->addCreditToFrame(credit);
         }
       }
+      if (tile->getContext()->implicitContext &&
+          tile->getContext()->implicitContext->credit) {
+        pCreditSystem->addCreditToFrame(
+            *tile->getContext()->implicitContext->credit);
+      }
     }
   }
 
@@ -389,15 +396,105 @@ void Tileset::loadTilesFromJson(
 
 CesiumAsync::Future<std::shared_ptr<CesiumAsync::IAssetRequest>>
 Tileset::requestTileContent(Tile& tile) {
-  std::string url = this->getResolvedContentUrl(tile);
+  std::string url =
+      this->getResolvedContentUrl(*tile.getContext(), tile.getTileID());
   assert(!url.empty());
 
   this->notifyTileStartLoading(&tile);
 
-  return this->getExternals().pAssetAccessor->get(
-      this->getAsyncSystem(),
-      url,
-      tile.getContext()->requestHeaders);
+  Future<std::shared_ptr<IAssetRequest>> tileFuture =
+      this->getExternals().pAssetAccessor->get(
+          this->getAsyncSystem(),
+          url,
+          tile.getContext()->requestHeaders);
+
+  // In addition to loading the main tile, we also need to load the same tile in
+  // any underlying contexts for which this tile is an availability level. This
+  // is necessary because, when we later create this tile's children, we need to
+  // be able to create children that are only available from an underlying
+  // layer, and we can only do that if we know they're available.
+  TileContext* pTileContext = tile.getContext();
+  TileContext* pContext =
+      pTileContext->pTopContext ? pTileContext->pTopContext : pTileContext;
+
+  const QuadtreeTileID* pQuadtreeTileID =
+      std::get_if<QuadtreeTileID>(&tile.getTileID());
+
+  if (pQuadtreeTileID && pContext->implicitContext &&
+      pContext->implicitContext->rectangleAvailability &&
+      pContext->implicitContext->availabilityLevels) {
+    uint32_t level = pQuadtreeTileID->level;
+    std::vector<Future<int>> layerFutures;
+    while (pContext) {
+      if (pContext != pTileContext && pContext->implicitContext &&
+          pContext->implicitContext->availabilityLevels) {
+        uint32_t availabilityLevels =
+            *pContext->implicitContext->availabilityLevels;
+        if ((level % availabilityLevels) == 0) {
+          // This tile has availability data, so load it.
+          layerFutures.emplace_back(this->_requestQuantizedMeshAvailabilityTile(
+              *pQuadtreeTileID,
+              pContext));
+        }
+      }
+
+      pContext = pContext->pUnderlyingContext.get();
+    }
+
+    // If we need to load any tiles for availability, do that and then return
+    // the data for the actual tile.
+    if (!layerFutures.empty()) {
+      return this->getAsyncSystem()
+          .all(std::move(layerFutures))
+          .thenImmediately(
+              [tileFuture = std::move(tileFuture)](std::vector<int>&&) mutable {
+                return std::move(tileFuture);
+              });
+    }
+  }
+
+  return tileFuture;
+}
+
+Future<int> Tileset::_requestQuantizedMeshAvailabilityTile(
+    const CesiumGeometry::QuadtreeTileID& availabilityTileID,
+    TileContext* pAvailabilityContext) {
+  std::string url =
+      getResolvedContentUrl(*pAvailabilityContext, availabilityTileID);
+
+  return this->getExternals()
+      .pAssetAccessor
+      ->get(this->getAsyncSystem(), url, pAvailabilityContext->requestHeaders)
+      .thenInWorkerThread(
+          [pLogger = _externals.pLogger,
+           availabilityTileID](std::shared_ptr<IAssetRequest>&& pRequest)
+              -> std::vector<QuadtreeTileRectangularRange> {
+            const IAssetResponse* pResponse = pRequest->response();
+            if (pResponse) {
+              uint16_t statusCode = pResponse->statusCode();
+
+              if (!(statusCode != 0 &&
+                    (statusCode < 200 || statusCode >= 300))) {
+                return QuantizedMeshContent::loadMetadata(
+                    pLogger,
+                    pResponse->data(),
+                    availabilityTileID);
+              }
+            }
+            return {};
+          })
+      .thenInMainThread(
+          [pAvailabilityContext](
+              std::vector<CesiumGeometry::QuadtreeTileRectangularRange>&&
+                  rectangles) {
+            if (!rectangles.empty()) {
+              for (const QuadtreeTileRectangularRange& range : rectangles) {
+                pAvailabilityContext->implicitContext->rectangleAvailability
+                    ->addAvailableTileRange(range);
+              }
+            }
+            return 0;
+          });
 }
 
 void Tileset::addContext(std::unique_ptr<TileContext>&& pNewContext) {
@@ -735,18 +832,21 @@ Tileset::TraversalDetails Tileset::_renderLeaf(
   return traversalDetails;
 }
 
-bool Tileset::_queueLoadOfChildrenRequiredForRefinement(
+bool Tileset::_queueLoadOfChildrenRequiredForForbidHoles(
     const FrameState& frameState,
     Tile& tile,
     const ImplicitTraversalInfo& implicitInfo,
     const std::vector<double>& distances) {
-  if (!this->_options.forbidHoles) {
-    return false;
-  }
+  // This method should only be called in "Forbid Holes" mode.
+  assert(this->_options.forbidHoles);
+
+  bool waitingForChildren = false;
+
   // If we're forbidding holes, don't refine if any children are still loading.
   gsl::span<Tile> children = tile.getChildren();
-  bool waitingForChildren = false;
   for (Tile& child : children) {
+    this->_markTileVisited(child);
+
     if (!child.isRenderable() && !child.isExternalTileset()) {
       waitingForChildren = true;
 
@@ -761,7 +861,6 @@ bool Tileset::_queueLoadOfChildrenRequiredForRefinement(
             childInfo);
       }
       child.update(frameState.lastFrameNumber, frameState.currentFrameNumber);
-      this->_markTileVisited(child);
 
       // We're using the distance to the parent tile to compute the load
       // priority. This is fine because the relative priority of the children is
@@ -772,8 +871,22 @@ bool Tileset::_queueLoadOfChildrenRequiredForRefinement(
           frameState.frustums,
           child,
           distances);
+    } else if (child.getUnconditionallyRefine()) {
+      // This child tile is set to unconditionally refine. That means refining
+      // _to_ it will immediately refine _through_ it. So we need to make sure
+      // its children are renderable, too.
+      // The distances are not correct for the child's children, but once again
+      // we don't care because all tiles must be loaded before we can render any
+      // of them, so their relative priority doesn't matter.
+      ImplicitTraversalInfo childInfo(&child, &implicitInfo);
+      waitingForChildren |= this->_queueLoadOfChildrenRequiredForForbidHoles(
+          frameState,
+          child,
+          childInfo,
+          distances);
     }
   }
+
   return waitingForChildren;
 }
 
@@ -959,6 +1072,8 @@ bool Tileset::_kickDescendantsAndRenderTile(
   // in that case, load this tile INSTEAD of loading any of the descendants, and
   // tell the up-level we're only waiting on this tile. Keep doing this until we
   // actually manage to render this tile.
+  // Make sure we don't end up waiting on a tile that will _never_ be
+  // renderable.
   const bool wasRenderedLastFrame =
       lastFrameSelectionState.getResult(frameState.lastFrameNumber) ==
       TileSelectionState::Result::Rendered;
@@ -967,7 +1082,8 @@ bool Tileset::_kickDescendantsAndRenderTile(
 
   if (!wasReallyRenderedLastFrame &&
       traversalDetails.notYetRenderableCount >
-          this->_options.loadingDescendantLimit) {
+          this->_options.loadingDescendantLimit &&
+      !tile.isExternalTileset() && !tile.getUnconditionallyRefine()) {
     // Remove all descendants from the load queues.
     this->_loadQueueLow.erase(
         this->_loadQueueLow.begin() +
@@ -1034,14 +1150,23 @@ Tileset::TraversalDetails Tileset::_visitTile(
 
   const bool unconditionallyRefine = tile.getUnconditionallyRefine();
   const bool meetsSse = _meetsSse(frameState.frustums, tile, distances, culled);
-  const bool waitingForChildren = _queueLoadOfChildrenRequiredForRefinement(
-      frameState,
-      tile,
-      implicitInfo,
-      distances);
 
-  if (!unconditionallyRefine &&
-      (meetsSse || ancestorMeetsSse || waitingForChildren)) {
+  bool wantToRefine = unconditionallyRefine || (!meetsSse && !ancestorMeetsSse);
+
+  // In "Forbid Holes" mode, we cannot refine this tile until all its children
+  // are loaded. But don't queue the children for load until we _want_ to
+  // refine this tile.
+  if (wantToRefine && this->_options.forbidHoles) {
+    const bool waitingForChildren =
+        this->_queueLoadOfChildrenRequiredForForbidHoles(
+            frameState,
+            tile,
+            implicitInfo,
+            distances);
+    wantToRefine = !waitingForChildren;
+  }
+
+  if (!wantToRefine) {
     // This tile (or an ancestor) is the one we want to render this frame, but
     // we'll do different things depending on the state of this tile and on what
     // we did _last_ frame.
@@ -1246,7 +1371,9 @@ void Tileset::_markTileVisited(Tile& tile) noexcept {
   this->_loadedTiles.insertAtTail(tile);
 }
 
-std::string Tileset::getResolvedContentUrl(const Tile& tile) const {
+std::string Tileset::getResolvedContentUrl(
+    const TileContext& context,
+    const TileID& tileID) const {
   struct Operation {
     const TileContext& context;
 
@@ -1311,12 +1438,12 @@ std::string Tileset::getResolvedContentUrl(const Tile& tile) const {
     }
   };
 
-  std::string url = std::visit(Operation{*tile.getContext()}, tile.getTileID());
+  std::string url = std::visit(Operation{context}, tileID);
   if (url.empty()) {
     return url;
   }
 
-  return CesiumUtility::Uri::resolve(tile.getContext()->baseUrl, url, true);
+  return CesiumUtility::Uri::resolve(context.baseUrl, url, true);
 }
 
 static bool anyRasterOverlaysNeedLoading(const Tile& tile) noexcept {
@@ -1406,6 +1533,38 @@ static bool anyRasterOverlaysNeedLoading(const Tile& tile) noexcept {
       // state if needed.
       if (tile.getState() == Tile::LoadState::Unloaded) {
         tile.setState(Tile::LoadState::ContentLoaded);
+
+        // There are two possible ways to handle a tile with no content:
+        //
+        // 1. Treat it as a placeholder used for more efficient culling, but
+        //    never render it. Refining to this tile is equivalent to refining
+        //    to its children. To have this behavior, the tile _should_ have
+        //    content, but that content's model should be std::nullopt.
+        // 2. Treat it as an indication that nothing need be rendered in this
+        //    area at this level-of-detail. In other words, "render" it as a
+        //    hole. To have this behavior, the tile should _not_ have content at
+        //    all.
+        //
+        // We distinguish whether the tileset creator wanted (1) or (2) by
+        // comparing this tile's geometricError to the geometricError of its
+        // parent tile. If this tile's error is greater than or equal to its
+        // parent, treat it as (1). If it's less, treat it as (2).
+        //
+        // For a tile with no parent there's no difference between the
+        // behaviors.
+        double myGeometricError = tile.getNonZeroGeometricError();
+
+        Tile* pAncestor = tile.getParent();
+        while (pAncestor && pAncestor->getUnconditionallyRefine()) {
+          pAncestor = pAncestor->getParent();
+        }
+
+        double parentGeometricError =
+            pAncestor ? pAncestor->getNonZeroGeometricError()
+                      : myGeometricError * 2.0;
+        if (myGeometricError >= parentGeometricError) {
+          tile.setEmptyContent();
+        }
       }
     } else if (shouldLoad) {
       loadQueue.push_back({&tile, highestLoadPriority});