+++ Cesium3DTilesSelection/test/TestQuantizedMeshContent.cpp
@@ -704,6 +704,8 @@ TEST_CASE("Test converting quantized mesh to gltf with skirt") {
       projection,
       QuadtreeRectangleAvailability(tilingScheme, 23),
       std::nullopt,
+      std::nullopt,
+      std::nullopt,
       std::nullopt};
 
   SECTION("Check quantized mesh that has uint16_t indices") {
@@ -1042,6 +1044,8 @@ TEST_CASE("Test converting ill-formed quantized mesh") {
       boundingRegion,
       projection,
       QuadtreeRectangleAvailability(tilingScheme, 23),
+      std::nullopt,
+      std::nullopt,
       std::nullopt,
       std::nullopt};